Removes as<other> methods from Collection.
Removes asList and asSet and asSequence from Collection.
This makes Collection and Iterable identical, so also renames Iterable
to Collection.  Changes all uses of the removed methods to uses of the
sequence, list, and set constructors.  Removes redundant tests.
Makes C tests as similar to the JS tests as possible.