Records the location of dialect and import statements.
Now, stringTokens store the size of the string in the input,
which will exceed the size of the string itself, because of
embedded escapes and quoting. In the AST, stringNodes
now record the end of the string, and therefor correctly
respond to the `range` observer.

Stores end rather than size of stringNodes.

That's all.